Bug 1865078 part 8 - Discard JIT code per realm instead of per zone. r=iain,jonco
`GCRuntime::maybeDiscardJitCodeForGC` is now the only place where the JitRealm's
preserve-code flags are set, because we clear the flags after discarding JIT code.
We assert the flags are cleared in `Zone::discardJitCodeForAllRealms`.

Only for realms that are not preserving code, we now invalidate scripts, mark
scripts as active, or purge the IC stub LifoAlloc.

Differential Revision: https://phabricator.services.mozilla.com/D326164

+++ js/src/gc/Zone.cpp
@@ -311,21 +311,40 @@ bool Zone::isAnyRealmPreservingCode() {
   return false;
 }
 
-void Zone::forceDiscardJitCode(JS::GCContext* gcx,
-                               const JitDiscardOptions& options) {
+void Zone::discardJitCodeForAllRealms(JS::GCContext* gcx) {
+  MOZ_ASSERT(!isAnyRealmPreservingCode());
+  discardJitCode(gcx);
+}
+
+void Zone::discardJitCode(JS::GCContext* gcx,
+                          const JitDiscardOptions& options) {
   if (!jitZone()) {
     return;
   }
 
-  // Move the IC stub data of each realm in this zone to a separate LifoAlloc.
-  // Stubs that must survive are copied back to their realm's stub space below.
+  // Move IC stub data from realms not preserving JIT code to a separate
+  // LifoAlloc. Stubs that must survive are copied back to their realm's stub
+  // space below.
   jit::ICStubSpace discardedStubSpace;
   const mozilla::TimeStamp now = mozilla::TimeStamp::Now();
+  size_t numRealms = 0;
+  size_t numDiscardedRealms = 0;
   for (RealmsInZoneIter r(this); !r.done(); r.next()) {
+    numRealms++;
+    if (r->jitRealm().isPreservingCode()) {
+      continue;
+    }
+    numDiscardedRealms++;
     if (options.discardJitScripts) {
       r->jitRealm().setLastDiscardedCodeTime(now);
     }
     discardedStubSpace.transferFrom(*r->jitRealm().stubSpace());
+  }
+
+  bool resetAllocSites =
+      options.resetNurseryAllocSites || options.resetPretenuredAllocSites;
+  if (numDiscardedRealms == 0 && !resetAllocSites) {
+    return;
   }
 
 #ifdef DEBUG
@@ -338,12 +357,28 @@ void Zone::forceDiscardJitCode(JS::GCContext* gcx,
   // Mark ICScripts on the stack as active and copy active Baseline stubs.
   jit::MarkActiveICScriptsAndCopyStubs(this);
 
-  // Invalidate all Ion code in this zone.
+  // Invalidate all Ion code in this zone for realms not preserving JIT code.
   jit::InvalidateAll(gcx, this);
 
   jitZone()->forEachJitScript<jit::IncludeDyingScripts>(
       [&](jit::JitScript* jitScript) {
         JSScript* script = jitScript->owningScript();
+        if (script->realm()->jitRealm().isPreservingCode()) {
+          // We're not discarding this realm's JIT code, but we may still have
+          // to reset allocation sites.
+          if (resetAllocSites &&
+              jitScript->resetAllocSites(options.resetNurseryAllocSites,
+                                         options.resetPretenuredAllocSites)) {
+            if (script->hasIonScript()) {
+              jit::Invalidate(runtime_->mainContextFromOwnThread(), script,
+                              /* resetUses = */ true,
+                              /* cancelOffThread = */ false);
+            }
+          }
+          MOZ_ASSERT(!jitScript->icScript()->active());
+          return;
+        }
+
         jit::FinishInvalidation(gcx, script);
 
         // Discard baseline script if it's not marked as active.
@@ -386,8 +421,7 @@ void Zone::forceDiscardJitCode(JS::GCContext* gcx,
         jitScript->purgeInactiveICScripts();
         jitScript->purgeStubs(script);
 
-        if (options.resetNurseryAllocSites ||
-            options.resetPretenuredAllocSites) {
+        if (resetAllocSites) {
           jitScript->resetAllocSites(options.resetNurseryAllocSites,
                                      options.resetPretenuredAllocSites);
         }
@@ -397,9 +431,12 @@ void Zone::forceDiscardJitCode(JS::GCContext* gcx,
       });
 
   // Also clear references to jit code from RegExpShared cells at this point.
-  // This avoid holding onto ExecutablePools. Do not discard if an interrupted
-  // regexp is currently on the stack.
-  if (!jitZone()->keepRegExpJitCode()) {
+  // This avoid holding onto ExecutablePools. RegExpShareds are not associated
+  // with a realm, so we only do this when discarding JIT code for all realms.
+  // Do not discard if an interrupted regexp is currently on the stack.
+  const bool discardRegExpJitCode =
+      numDiscardedRealms == numRealms && !jitZone()->keepRegExpJitCode();
+  if (discardRegExpJitCode) {
     for (auto regExp = cellIterUnsafe<RegExpShared>(); !regExp.done();
          regExp.next()) {
       regExp->discardJitCode();
@@ -416,7 +453,9 @@ void Zone::forceDiscardJitCode(JS::GCContext* gcx,
    */
   discardedStubSpace.freeAllAfterMinorGC(this);
   for (RealmsInZoneIter r(this); !r.done(); r.next()) {
-    r->jitRealm().purgeIonCacheIRStubInfo();
+    if (!r->jitRealm().isPreservingCode()) {
+      r->jitRealm().purgeIonCacheIRStubInfo();
+    }
   }
 
   // Generate a profile marker
@@ -425,44 +464,20 @@ void Zone::forceDiscardJitCode(JS::GCContext* gcx,
     char discardingBaseline = 'Y';
     char discardingIon = 'Y';
 
-    char discardingRegExp = jitZone()->keepRegExpJitCode() ? 'N' : 'Y';
+    char discardingRegExp = discardRegExpJitCode ? 'Y' : 'N';
     char discardingNurserySites = options.resetNurseryAllocSites ? 'Y' : 'N';
     char discardingPretenuredSites =
         options.resetPretenuredAllocSites ? 'Y' : 'N';
 
-    char buf[100];
+    char buf[160];
     SprintfLiteral(buf,
-                   "JitScript:%c Baseline:%c Ion:%c "
+                   "Realms:%zu/%zu JitScript:%c Baseline:%c Ion:%c "
                    "RegExp:%c NurserySites:%c PretenuredSites:%c",
-                   discardingJitScript, discardingBaseline, discardingIon,
-                   discardingRegExp, discardingNurserySites,
-                   discardingPretenuredSites);
+                   numDiscardedRealms, numRealms, discardingJitScript,
+                   discardingBaseline, discardingIon, discardingRegExp,
+                   discardingNurserySites, discardingPretenuredSites);
     gcx->runtime()->geckoProfiler().markEvent("DiscardJit", buf);
   }
-}
-
-void JS::Zone::resetAllocSitesAndInvalidate(bool resetNurserySites,
-                                            bool resetPretenuredSites) {
-  MOZ_ASSERT(resetNurserySites || resetPretenuredSites);
-
-  if (!jitZone()) {
-    return;
-  }
-
-  JSContext* cx = runtime_->mainContextFromOwnThread();
-  jitZone()->forEachJitScript<jit::IncludeDyingScripts>(
-      [&](jit::JitScript* jitScript) {
-        if (jitScript->resetAllocSites(resetNurserySites,
-                                       resetPretenuredSites)) {
-          JSScript* script = jitScript->owningScript();
-          CancelOffThreadIonCompile(script);
-          if (script->hasIonScript()) {
-            jit::Invalidate(cx, script,
-                            /* resetUses = */ true,
-                            /* cancelOffThread = */ true);
-          }
-        }
-      });
 }
 
 void JS::Zone::traceWeakJitScripts(JSTracer* trc) {
@@ -543,9 +558,7 @@ Zone* Zone::nextZone() const {
 
 void Zone::prepareForMovingGC() {
   JS::GCContext* gcx = runtimeFromMainThread()->gcContext();
-
-  MOZ_ASSERT(!isAnyRealmPreservingCode());
-  forceDiscardJitCode(gcx);
+  discardJitCodeForAllRealms(gcx);
 }
 
 void Zone::fixupAfterMovingGC() {
